Fix: name or value nil
When `cookie-name` is `nil` or an empty string `""`,
an `invalid-cookie` error is signaled.

If `*sanity-check*` is `nil`, `write-cookie-header`
and `write-set-coookie-header` do not print anything as `cookie-name`.
Till now they were printing `NIL`.

`write-cookie-header` and `write-set-coookie-header` do not
print anything as `cookie-value` if not supplied.
Till now they were printing `NIL`.